backend/requirements.txt
# VegaExchange Backend Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
asyncpg>=0.29.0

# Data validation
pydantic>=2.5.0
email-validator>=2.0.0

# Environment
python-dotenv>=1.0.0

# Authentication (for Google OAuth and email/password)
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
passlib[bcrypt]>=1.7.4

# Utilities
python-multipart>=0.0.6
orjson>=3.9.0
